cleanup redundant use client, convert portfolio page to server comp
