+++ app/ui/NavBar/NavButton.tsx
@@ -1,5 +1,4 @@
 "use client"
-
 import React from 'react';
 import clsx from "clsx";
 import Link from "next/link";
